# level2/task2/docs/swagger.yaml
swagger: "2.0"
basePath: /api/v1
definitions:
...
  controller.ContractBytecodeRequest:
    properties:
      contract_type:
        example: SimpleStorage
        type: string
    required:
    - contract_type
    type: object
  controller.ContractBytecodeResponse:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      bytecode:
        example: 0x608060405234801561001057600080fd5b50610150806100206000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c80632e64cec11461003b5780636057361d14610059575b600080fd5b610043610075565b60405161005091906100d9565b60405180910390f35b610073600480360381019061006e919061009d565b61007e565b005b60008054905090565b8060008190555050565b60008135905061009781610103565b92915050565b6000602082840312156100b3576100b26100fe565b5b60006100c184828501610088565b91505092915050565b6100d3816100f4565b82525050565b60006020820190506100ee60008301846100ca565b92915050565b6000819050919050565b600080fd5b61010c816100f456b811461011757600080fd5b5056fea26469706673582212209a159a4f3847890f10d0e3a00307a5b6bc6608c2f89fd2d8bd8d7e72a7c1d7d064736f6c63430008070033
        type: string
      contract_type:
        example: SimpleStorage
        type: string
    type: object
  controller.ContractResponse:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      contract_type:
        example: SimpleStorage
        type: string
      error:
        example: 交易失败
        type: string
      txHash:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
    type: object
  controller.ErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: 服务器内部错误
        type: string
    type: object
  controller.LockWithdrawRequest:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 提供以太坊区块链相关的API服务，包括账户管理、交易处理、合约部署等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Web3 区块链接口服务
  version: "1.0"
paths:
...
  /contract/lock/withdraw:
    post:
      consumes:
      - application/json
      description: 调用Lock合约的Withdraw方法，提取合约中的资金
      parameters:
      - description: 提取资金请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.LockWithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ContractResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 执行Lock合约的Withdraw方法
      tags:
      - 锁定合约
...
  /contracts/allAddresses:
    get:
      consumes:
      - application/json
      description: 获取所有合约地址，包括SimpleStorage、SimpleAuction、Purchase、Lottery等智能合约
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取所有合约地址
      tags:
      - 合约
  /contracts/bytecode:
    post:
      consumes:
      - application/json
      description: 获取指定合约的字节码
      parameters:
      - description: 合约类型和合约地址
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ContractBytecodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ContractBytecodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取合约字节码
      tags:
      - 合约
  /contracts/deploy-all:
    post:
      consumes:
      - application/json
      description: 部署所有合约，包括SimpleStorage、SimpleAuction、Purchase、Lottery等智能合约
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 部署所有合约
      tags:
      - 合约
  /events/subscribe:
    get:
      consumes:
      - application/json
      description: 通过WebSocket连接订阅合约事件，实时接收事件数据
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols to websocket
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 订阅合约事件
      tags:
      - 事件
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey